@@ -7,16 +7,22 @@
         }
     },
     "require": {
-        "php": ">=5.5",
-        "symfony/symfony": "^2.6|3.*",
-        "paysera/lib-serializer": "^2.0.2",
-        "doctrine/orm": "^2.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^4.8",
-        "mockery/mockery": "^0.9.4"
+        "php": "^7.0",
+        "paysera/lib-serializer": "^3.0",
+        "doctrine/orm": "^2.0",
+        "symfony/config": "^3.0 || ^4.0",
+        "symfony/dependency-injection": "^3.0 || ^4.0",
+        "symfony/validator": "^3.0 || ^4.0",
+        "symfony/security-core": "^3.0 || ^4.0",
+        "symfony/routing": "^3.0 || ^4.0",
+        "symfony/http-foundation": "^3.0 || ^4.0",
+        "symfony/http-kernel": "^3.0 || ^4.0"
     },
     "config": {
         "bin-dir": "bin"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^6.5",
+        "mockery/mockery": "^1.2"
     }
 }
