@@ -2,6 +2,8 @@
 
 namespace Paysera\Bundle\RestBundle\Cache;
 
+use DateTime;
+
 class DefaultCacheStrategy implements CacheStrategyInterface
 {
     protected $maxAge;
@@ -24,7 +26,7 @@ class DefaultCacheStrategy implements CacheStrategyInterface
     /**
      * @param mixed $result Returned from controller, not normalizer
      *
-     * @return \DateTime|null
+     * @return DateTime|null
      */
     public function getModifiedAt($result)
     {
