@@ -2,6 +2,7 @@
 
 namespace Paysera\Bundle\RestBundle\Listener;
 
+use Exception;
 use Paysera\Bundle\RestBundle\Cache\ResponseAwareCacheStrategy;
 use Paysera\Bundle\RestBundle\Service\ExceptionLogger;
 use Paysera\Bundle\RestBundle\Service\ParameterToEntityMapBuilder;
@@ -90,7 +91,7 @@ class RestListener
      *
      * @throws ApiException
      * @throws InvalidDataException
-     * @throws \Exception
+     * @throws Exception
      */
     public function onKernelController(FilterControllerEvent $event)
     {
@@ -310,12 +311,16 @@ class RestListener
                     );
                 }
             }
+            $entity = null;
             try {
                 $entity = $requestMapper->mapToEntity($data);
             } catch (InvalidDataException $exception) {
                 $this->handleException($exception);
             }
-            $this->getLogger($request)->debug('Mapped data to entity', array($entity));
+            $this->getLogger($request)->debug(
+                'Mapped data to entity',
+                ['entity' => $entity]
+            );
 
             $this->validateEntity($request, $entity);
 
@@ -334,12 +339,15 @@ class RestListener
     {
         $requestQueryMapper = $this->apiManager->getRequestQueryMapper($request);
         if ($requestQueryMapper !== null) {
+            $entity = null;
             try {
                 $entity = $requestQueryMapper->mapToEntity($request->query->all());
             } catch (InvalidDataException $exception) {
                 $this->handleException($exception);
             }
-            $this->getLogger($request)->debug('Mapped query data to entity', array($entity));
+            $this->getLogger($request)->debug('Mapped query data to entity', [
+                'entity' => $entity,
+            ]);
 
             $this->validateEntity($request, $entity);
 
