@@ -2,6 +2,7 @@
 
 namespace Paysera\Bundle\RestBundle\ModificationDateProvider;
 
+use DateTime;
 use Paysera\Bundle\RestBundle\Cache\ModificationDateProviderInterface;
 
 class CollectionDateProvider implements ModificationDateProviderInterface
@@ -16,7 +17,7 @@ class CollectionDateProvider implements ModificationDateProviderInterface
     /**
      * @param array $items
      *
-     * @return \DateTime|null
+     * @return DateTime|null
      */
     public function getModifiedAt($items)
     {
