@@ -2,15 +2,20 @@
 
 namespace Paysera\Bundle\RestBundle;
 
+use Closure;
 use Paysera\Bundle\RestBundle\Cache\CacheStrategyInterface;
+use Paysera\Bundle\RestBundle\Normalizer\NameAwareDenormalizerInterface;
 use Paysera\Bundle\RestBundle\Resolver\EntityResolverInterface;
 use Paysera\Bundle\RestBundle\Resolver\ServiceAwareAttributeResolver;
+use Paysera\Component\Serializer\Encoding\DecoderInterface;
+use Paysera\Component\Serializer\Encoding\EncoderInterface;
 use Paysera\Component\Serializer\Factory\EncoderFactoryInterface;
 use Paysera\Component\Serializer\Factory\ResponseMapperFactoryInterface;
 use Paysera\Bundle\RestBundle\Normalizer\NameAwareDenormalizer;
 use Paysera\Component\Serializer\Normalizer\DenormalizerInterface;
 use Paysera\Component\Serializer\Validation\PropertyPathConverterInterface;
 use Psr\Log\LoggerInterface;
+use RuntimeException;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Paysera\Bundle\RestBundle\Entity\ErrorConfig;
 use Paysera\Bundle\RestBundle\Security\SecurityStrategyInterface;
@@ -21,17 +26,17 @@ class RestApi
     const DEFAULT_VALIDATION_GROUP = 'Default';
 
     /**
-     * @var \Closure[]
+     * @var Closure[]
      */
     protected $requestMappers;
 
     /**
-     * @var \Closure[]
+     * @var Closure[]
      */
     protected $requestQueryMappers;
 
     /**
-     * @var \Closure[]
+     * @var Closure[]
      */
     protected $requestAttributeResolvers;
 
@@ -46,7 +51,7 @@ class RestApi
     protected $responseMappers;
 
     /**
-     * @var \Paysera\Component\Serializer\Factory\ResponseMapperFactoryInterface[]
+     * @var ResponseMapperFactoryInterface[]
      */
     protected $responseMapperFactories;
 
@@ -81,7 +86,7 @@ class RestApi
     protected $securityStrategy;
 
     /**
-     * @var \Symfony\Component\DependencyInjection\ContainerInterface
+     * @var ContainerInterface
      */
     protected $serviceContainer;
 
@@ -113,7 +118,7 @@ class RestApi
     /**
      * Constructs object
      *
-     * @param \Symfony\Component\DependencyInjection\ContainerInterface $serviceContainer
+     * @param ContainerInterface $serviceContainer
      * @param LoggerInterface                                           $logger
      */
     public function __construct(
@@ -181,7 +186,7 @@ class RestApi
      * @param string $mapperKey    Denormalizer mapper key
      * @param string $argumentName Denormalizer name
      *
-     * @return \Closure
+     * @return Closure
      */
     protected function getDenormalizerClosure($mapperKey, $argumentName)
     {
@@ -190,7 +195,7 @@ class RestApi
         return function() use ($mapperKey, $serviceContainer, $argumentName) {
             $denormalizer = $serviceContainer->get($mapperKey);
             if (!$denormalizer instanceof DenormalizerInterface) {
-                throw new \RuntimeException(
+                throw new RuntimeException(
                     'Configured service does not implement DenormalizerInterface ' . $mapperKey
                 );
             }
@@ -273,7 +278,7 @@ class RestApi
             /** @var EntityResolverInterface $entityResolverService */
             $implementsResolverInterface = $entityResolverService instanceof EntityResolverInterface;
             if (!$implementsResolverInterface) {
-                throw new \RuntimeException(
+                throw new RuntimeException(
                     'The entity resolver service needs to implement EntityResolverInterface'
                 );
             }
@@ -297,7 +302,7 @@ class RestApi
     /**
      * Adds response mapper for some controller
      *
-     * @param \Paysera\Component\Serializer\Factory\ResponseMapperFactoryInterface $mapperFactory
+     * @param ResponseMapperFactoryInterface $mapperFactory
      * @param string                                                           $controllerKey
      */
     public function addResponseMapperFactory(ResponseMapperFactoryInterface $mapperFactory, $controllerKey)
@@ -327,7 +332,7 @@ class RestApi
      * Adds supported response format
      *
      * @param string                                                    $format
-     * @param \Paysera\Component\Serializer\Factory\EncoderFactoryInterface $encoderFactory
+     * @param EncoderFactoryInterface $encoderFactory
      */
     public function addResponseEncoderFactory($format, EncoderFactoryInterface $encoderFactory)
     {
@@ -356,7 +361,7 @@ class RestApi
     /**
      * Sets errorConfig
      *
-     * @param \Paysera\Bundle\RestBundle\Entity\ErrorConfig $errorConfig
+     * @param ErrorConfig $errorConfig
      */
     public function setErrorConfig($errorConfig)
     {
@@ -366,7 +371,7 @@ class RestApi
     /**
      * Sets securityStrategy
      *
-     * @param \Paysera\Bundle\RestBundle\Security\SecurityStrategyInterface|null $securityStrategy
+     * @param SecurityStrategyInterface|null $securityStrategy
      */
     public function setSecurityStrategy($securityStrategy)
     {
@@ -429,7 +434,7 @@ class RestApi
      *
      * @param string $controllerKey
      *
-     * @return \Paysera\Bundle\RestBundle\Normalizer\NameAwareDenormalizerInterface|null
+     * @return NameAwareDenormalizerInterface|null
      */
     public function getRequestMapper($controllerKey)
     {
@@ -448,7 +453,7 @@ class RestApi
      *
      * @param string $controllerKey
      *
-     * @return \Paysera\Bundle\RestBundle\Normalizer\NameAwareDenormalizerInterface|null
+     * @return NameAwareDenormalizerInterface|null
      */
     public function getRequestQueryMapper($controllerKey)
     {
@@ -488,13 +493,14 @@ class RestApi
      *
      * @param string $controllerKey
      *
-     * @return \Paysera\Bundle\RestBundle\Normalizer\NameAwareDenormalizerInterface|null|array
+     * @return NameAwareDenormalizerInterface|null|array
      */
     public function getRequestAttributeResolvers($controllerKey)
     {
         $this->logger->debug('Getting request attribute mapper for ' . $controllerKey);
         $controllerKey = $this->normalizeControllerKey($controllerKey);
         if (isset($this->requestAttributeResolvers[$controllerKey])) {
+            /** @var Closure[] $closures */
             $closures = $this->requestAttributeResolvers[$controllerKey];
             return array_map(
                 function ($closure) {
@@ -533,7 +539,7 @@ class RestApi
      * @param string $controllerKey
      * @param array  $options
      *
-     * @return \Paysera\Bundle\RestBundle\Cache\CacheStrategyInterface|null
+     * @return CacheStrategyInterface|null
      */
     public function getCacheStrategy($controllerKey, array $options = array())
     {
@@ -573,7 +579,7 @@ class RestApi
      *
      * @param string $format
      *
-     * @return \Paysera\Component\Serializer\Encoding\DecoderInterface
+     * @return DecoderInterface
      */
     public function getDecoder($format)
     {
@@ -586,7 +592,7 @@ class RestApi
      * @param string $format
      * @param array  $options
      *
-     * @return \Paysera\Component\Serializer\Encoding\EncoderInterface
+     * @return EncoderInterface
      */
     public function getEncoder($format, array $options = array())
     {
