@@ -2,11 +2,19 @@
 
 namespace Tests;
 
+use Mockery;
+use Mockery\MockInterface;
 use Paysera\Bundle\RestBundle\RestApi;
 use Paysera\Component\Serializer\Converter\CamelCaseToSnakeCaseConverter;
+use PHPUnit\Framework\TestCase;
+use RuntimeException;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
-class RestApiTest extends \PHPUnit_Framework_TestCase
+class RestApiTest extends TestCase
 {
+    /**
+     * @var MockInterface|ContainerInterface
+     */
     private $serviceContainer;
     private $logger;
     private $messageStorage = [];
@@ -14,7 +22,7 @@ class RestApiTest extends \PHPUnit_Framework_TestCase
     
     public function setUp()
     {
-        $this->serviceContainer = \Mockery::mock('Symfony\Component\DependencyInjection\ContainerInterface');
+        $this->serviceContainer = Mockery::mock('Symfony\Component\DependencyInjection\ContainerInterface');
 
         $this->serviceContainer
             ->shouldReceive('get')
@@ -22,7 +30,7 @@ class RestApiTest extends \PHPUnit_Framework_TestCase
             ->andReturn(new CamelCaseToSnakeCaseConverter())
         ;
 
-        $this->logger = \Mockery::mock('Psr\Log\LoggerInterface');
+        $this->logger = Mockery::mock('Psr\Log\LoggerInterface');
         $this->logger->shouldReceive('debug')->andReturnUsing($this->storeMessage());
     }
     
@@ -43,6 +51,8 @@ class RestApiTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @dataProvider controllerKeyProvider
+     *
+     * @param string $controllerKey
      */
     public function testNormalizeControllerKeyUsingGetValidationGroups($controllerKey)
     {
@@ -73,7 +83,7 @@ class RestApiTest extends \PHPUnit_Framework_TestCase
     {
         $mapperKey = 'key';
         $argument = 'argument';
-        $mockedDenormalizer = \Mockery::mock('Paysera\Component\Serializer\Normalizer\DenormalizerInterface');
+        $mockedDenormalizer = Mockery::mock('Paysera\Component\Serializer\Normalizer\DenormalizerInterface');
         $this->serviceContainer->shouldReceive('get')->with($mapperKey)->andReturn($mockedDenormalizer);
 
         $restApi = new RestApi($this->serviceContainer, $this->logger);
@@ -92,9 +102,9 @@ class RestApiTest extends \PHPUnit_Framework_TestCase
 
     public function testSetRequestMapperThrowsExceptionWhenDenormalizerIsIncorrect()
     {
-        $this->setExpectedException('\RuntimeException');
+        $this->expectException(RuntimeException::class);
         $mapperKey = 'key';
-        $mockedDenormalizer = \Mockery::mock('RandomalizerInterface');
+        $mockedDenormalizer = Mockery::mock('RandomalizerInterface');
         $this->serviceContainer->shouldReceive('get')->with($mapperKey)->andReturn($mockedDenormalizer);
         $restApi = new RestApi($this->serviceContainer, $this->logger);
         $restApi->addRequestMapper($mapperKey, $this->controllerKey, 'argument');
@@ -103,9 +113,9 @@ class RestApiTest extends \PHPUnit_Framework_TestCase
 
     public function testSetRequestQueryMapperThrowsExceptionWhenDenormalizerIsIncorrect()
     {
-        $this->setExpectedException('\RuntimeException');
+        $this->expectException(RuntimeException::class);
         $mapperKey = 'key';
-        $mockedDenormalizer = \Mockery::mock('RandomalizerInterface');
+        $mockedDenormalizer = Mockery::mock('RandomalizerInterface');
         $this->serviceContainer->shouldReceive('get')->with($mapperKey)->andReturn($mockedDenormalizer);
         $restApi = new RestApi($this->serviceContainer, $this->logger);
         $restApi->addRequestQueryMapper($mapperKey, $this->controllerKey, 'argument');
